Making rendering of NavBar depend on user session
If the user is logged in, then the nav bar components are that of a
logged in user. Otherwise, they contain paths to the login and signup
page. The Apollo cache still needs to be updated to automatically make
the update once the user logs in instead of requiring a manual refresh.